+++ src/js/jquery.form.generator.js
@@ -38,6 +38,20 @@
     }
 
     /**
+     * @param {String} expr
+     * @return {RegExp|null}
+     */
+    function expressionToRegex(expr) {
+        // for valid flags see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions#Advanced_searching_with_flags
+        var matches = expr.match(/^[/](.*?)[/]([gimsuy]*)$/);
+        if (matches) {
+            return new RegExp(matches[1], matches[2]);
+        } else {
+            return null;
+        }
+    }
+
+    /**
      * Check if typeof object[key] !== 'undefined'
      *
      * @param obj
@@ -253,7 +267,7 @@
                 // @todo: fold very apparent copy & paste between this method and "checkbox" method
                 var inputField = input;
                 if (!input) {
-                    var type = item.type || 'text';
+                    var type = (item.type !== 'input' && item.type) || 'text';
                     inputField = $('<input class="form-control" type="' + type + '"/>');
                 }
                 var container = $('<div class="form-group"/>');
@@ -281,29 +295,33 @@
                     container.addClass('has-title')
                 }
 
-                if(has(item, 'mandatory') && item.mandatory) {
+                if (item.mandatory) {
+                    var validationCallback;
+                    if (item.mandatory === true) {
+                        validationCallback = function(value) {
+                            return $.trim(value).length;
+                        };
+                    } else if (typeof item.mandatory === 'string') {
+                        // legacy fun fact: string runs through eval, but result of eval can only be used
+                        // if it happens to have an method named .exec accepting a single parameter
+                        // => this was never compatible with anything but regex literals
+                        var rxp = expressionToRegex(item.mandatory);
+                        if (rxp) {
+                            validationCallback = function(value) {
+                                return rxp.test(value);
+                            }
+                        }
+                    }
+                    if (!validationCallback) {
+                        console.error("Invalid value in item.mandatory. Use boolean true or a regex literal.", item.mandatory, item);
+                        throw new Error("Invalid value in item.mandatory. Use boolean true or a regex literal.");
+                    }
                     // @todo: why in the world is this a data attribute? Validation belongs in a form submit handler.
                     //        HTML5 validation already does most of this without custom logic
-                    inputField.data('warn',function(value){
-                        var hasValue = $.trim(value) != '';
-                        var isRegExp = item.mandatory !== true;
-
-                        if(isRegExp){
-                            console.error("Using Javascript code in the configuration is deprecated, but regular expression is ok",item.mandatory);
-                            hasValue = new RegExp(item.mandatory).exec(value) != null;
-                        }
-
-                        if(hasValue){
-                            container.removeClass('has-error');
-                        }else{
-                            if(inputField.is(":visible")){
-                                var text = item.hasOwnProperty('mandatoryText')? item.mandatoryText: "Please, check!";
-                                $.notify( inputField, text, { position:"top right", autoHideDelay: 2000});
-                            }
-                            container.addClass('has-error');
-                        }
-                        return hasValue;
-                    });
+                    inputField.data('warn', validationCallback);
+                }
+                if (item.mandatoryText) {
+                    inputField.attr('data-custom-error-message', item.mandatoryText);
                 }
 
                 if(has(item, 'infoText')) {
@@ -375,19 +393,8 @@
                 if(has(item, 'mandatory') && item.mandatory) {
                     // @todo: why in the world is this a data attribute? Validation belongs in a form submit handler.
                     //        HTML5 validation already does most of this without custom logic
-                    input.data('warn',function(){
-                        var isChecked = input.is(':checked');
-                        if(isChecked){
-                            container.removeClass('has-error');
-                        }else{
-                            container.addClass('has-error');
-                            if(input.is(':visible')){
-                                var text = item.hasOwnProperty('mandatoryText') ? item.mandatoryText : "Please confirm!";
-                                $.notify( input, text, { position:"top left", autoHideDelay: 2000});
-                            }
-
-                        }
-                        return isChecked;
+                    input.data('warn', function() {
+                        return input.is('checked');
                     });
                 }
 
@@ -498,7 +505,6 @@
                 if ((item.multiple || item.select2) && (typeof select.select2 === 'function')) {
                     select.select2(item);
                 }
-
 
                 return container;
             },